+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+
+## 5.0.1
+
+* Fix MacOS localized timezone name retrieval on OSX 13+
+
+## 5.0.0
+
+* Update minimum MacOS version to 10.14 ([50](https://github.com/tjarvstrand/flutter_timezone/pull/50)) @bqubique
+* Change API return values to TimezoneInfo objects containing both the timezone name and, on supported platforms, the localized timezone name.
+
+## 4.1.1
+
+* Don't always return true on Linux.
+
 ## 4.1.0
 
 * flutter_timezone now supports Swift Package Manager ([40](https://github.com/tjarvstrand/flutter_timezone/pull/44)) @MaikuB